Update README.md
Fixed markdown formatting issue with the use of underscores, that was hiding the proper underscore in the APPENGINE_HOME env. variable. Before it was italicizing the APPENGINE part and showing an underscore after HOME, now it atleast shows proper letters if not italicized. (Appears you can't italicize a work with an underscore in it in github markdown.